+++ NEWS.md
@@ -2,8 +2,8 @@
 
   * Updated to h3-js 3.6.1
   * Added `get_centerchild()`, which will return the central child of a hexagon at a given resolution.
-  * Added `get_pentagons()`, which will return the twelve pentagon indexes at a given resolution.  
-
+  * Added `get_pentagons()`, which will return the twelve pentagon indexes at a given resolution.
+  
 # v. 1.0
 
   * Updated to h3-js 3.5.0, now using official browser bundle